% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workHORSEmisc.R
\docType{package}
\name{workHORSEmisc}
\alias{workHORSEmisc-package}
\alias{workHORSEmisc}
\title{workHORSEmisc: A package with functions for the workHORSE model.}
\description{
Description of your package.
}
\author{
Chris Kypridemos
}
